% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{chart_cumgrowth}
\alias{chart_cumgrowth}
\title{Cumulative Growth Chart}
\usage{
chart_cumgrowth(
  ...,
  data = "prices",
  field = "adjusted",
  fieldname = "symbol",
  base = 100,
  name = NULL,
  title = "Cumulative Growth Chart",
  subtitle = "dates",
  timescale = "alpha",
  labels = "dates",
  pts.size = 0.1,
  pts.color = "black"
)
}
\arguments{
\item{pts.color}{}
}
\description{
Cumulative Growth Chart
}
