% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_famafrench}
\alias{get_famafrench}
\title{Get Fama-French Data}
\usage{
get_famafrench(dataset = "Fama/French 5 Factors (2x3)")
}
\arguments{
\item{dataset}{'Fama/French 5 Factors (2x3)' or 'Fama/French 3 Factors'}
}
\description{
Get Fama-French Data
}
\details{
"Fama/French 3 Factors"
"Fama/French 3 Factors [Weekly]"
"Fama/French 3 Factors [Daily]"
"Fama/French 5 Factors (2x3)"
"Fama/French 5 Factors (2x3) [Daily]"
"Portfolios Formed on Size"
"Portfolios Formed on Size [ex.Dividends]"
"Portfolios Formed on Size [Daily]"
"Portfolios Formed on Book-to-Market"
"Portfolios Formed on Book-to-Market [ex. Dividends]"
"Portfolios Formed on Book-to-Market [Daily]"
"Portfolios Formed on Operating Profitability"
"Portfolios Formed on Operating Profitability [ex. Dividends]"
"Portfolios Formed on Operating Profitability [Daily]"
"Portfolios Formed on Investment"
"Portfolios Formed on Investment [ex. Dividends]"
"Portfolios Formed on Investment [Daily]"
"6 Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 Portfolios Formed on Size and Book-to-Market (2 x 3) [ex. Dividends]"
"6 Portfolios Formed on Size and Book-to-Market (2 x 3) [Weekly]"
"6 Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"25 Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 Portfolios Formed on Size and Book-to-Market (5 x 5) [ex. Dividends]"
"25 Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"100 Portfolios Formed on Size and Book-to-Market (10 x 10)"
"100 Portfolios Formed on Size and Book-to-Market (10 x 10) [ex. Dividends]"
"100 Portfolios Formed on Size and Book-to-Market (10 x 10) [Daily]"
"6 Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 Portfolios Formed on Size and Operating Profitability (2 x 3) [ex. Dividends]"
"6 Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"25 Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 Portfolios Formed on Size and Operating Profitability (5 x 5) [ex. Dividends]"
"25 Portfolios Formed on Size and Operating Profitability	(5 x 5) [Daily]"
"100 Portfolios Formed on Size and Operating Profitability (10 x 10)"
"100 Portfolios Formed on Size and Operating Profitability (10 x 10) [ex. Dividends]"
"100 Portfolios Formed on Size and Operating Profitability (10 x 10) [Daily]"
"6 Portfolios Formed on Size and Investment (2 x 3)"
"6 Portfolios Formed on Size and Investment (2 x 3) [ex. Dividends]"
"6 Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"25 Portfolios Formed on Size and Investment (5 x 5)"
"25 Portfolios Formed on Size and Investment (5 x 5) [ex. Dividends]"
"25 Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"100 Portfolios Formed on Size and Investment (10 x 10)"
"100 Portfolios Formed on Size and Investment (10 x 10) [ex. Dividends]"
"100 Portfolios Formed on Size and Investment (10 x 10) [Daily]"
"25 Portfolios Formed on Book-to-Market and Operating Profitability (5 x 5)"
"25 Portfolios Formed on Book-to-Market and Operating Profitability (5 x 5) [ex. Dividends]"
"25 Portfolios Formed on Book-to-Market and Operating Profitability (5 x 5) [Daily]"
"25 Portfolios Formed on Book-to-Market and Investment (5 x 5)"
"25 Portfolios Formed on Book-to-Market and Investment (5 x 5) [ex. Dividends]"
"25 Portfolios Formed on Book-to-Market and Investment (5 x 5) [Daily]"
"25 Portfolios Formed on Operating Profitability and Investment (5 x 5)"
"25 Portfolios Formed on Operating Profitability and Investment (5 x 5) [ex. Dividends]"
"25 Portfolios Formed on Operating Profitability and Investment (5 x 5) [Daily]"
"32 Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4) [ex. Dividends]"
"32 Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4) [ex. Dividends]"
"32 Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"32 Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4) [ex. Dividends]"
"Portfolios Formed on Earnings/Price"
"Portfolios Formed on Earnings/Price [ex. Dividends]"
"Portfolios Formed on Cashflow/Price"
"Portfolios Formed on Cashflow/Price [ex. Dividends]"
"Portfolios Formed on Dividend Yield"
"Portfolios Formed on Dividend Yield [ex. Dividends]"
"6 Portfolios Formed on Size and Earnings/Price"
"6 Portfolios Formed on Size and Earnings/Price [ex. Dividends]"
"6 Portfolios Formed on Size and Cashflow/Price"
"6 Portfolios Formed on Size and Cashflow/Price [ex. Dividends]"
"6 Portfolios Formed on Size and Dividend Yield"
"6 Portfolios Formed on Size and Dividend Yield [ex. Dividends]"
"Momentum Factor (Mom)"
"Momentum Factor (Mom) [Daily]"
"6 Portfolios Formed on Size and Momentum (2 x 3)"
"6 Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"25 Portfolios Formed on Size and Momentum (5 x 5)"
"25 Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"10 Portfolios Formed on Momentum"
"10 Portfolios Formed on Momentum [Daily]"
"Short-Term Reversal Factor (ST Rev)"
"Short-Term Reversal Factor (ST Rev) [Daily]"
"6 Portfolios Formed on Size and Short-Term Reversal (2 x 3)"
"6 Portfolios Formed on Size and Short-Term Reversal (2 x 3) [Daily]"
"25 Portfolios Formed on Size and Short-Term Reversal (5 x 5)"
"25 Portfolios Formed on Size and Short-Term Reversal (5 x 5) [Daily]"
"10 Portfolios Formed on Short-Term Reversal"
"10 Portfolios Formed on Short-Term Reversal [Daily]"
"Long-Term Reversal Factor (LT Rev)"
"Long-Term Reversal Factor (LT Rev) [Daily]"
"6 Portfolios Formed on Size and Long-Term Reversal (2 x 3)"
"6 Portfolios Formed on Size and Long-Term Reversal (2 x 3) [Daily]"
"25 Portfolios Formed on Size and Long-Term Reversal (5 x 5)"
"25 Portfolios Formed on Size and Long-Term Reversal (5 x 5) [Daily]"
"10 Portfolios Formed on Long-Term Reversal"
"10 Portfolios Formed on Long-Term Reversal [Daily]"
"Portfolios Formed on Accruals"
"25 Portfolios Formed on Size and Accruals"
"Portfolios Formed on Market Beta"
"25 Portfolios Formed on Size and Market Beta"
"Portfolios Formed on Net Share Issues"
"25 Portfolios Formed on Size and Net Share Issues"
"Portfolios Formed on Variance"
"25 Portfolios Formed on Size and Variance"
"Portfolios Formed on Residual Variance"
"25 Portfolios Formed on Size and Residual Variance"
"5 Industry Portfolios"
"5 Industry Portfolios [ex. Dividends]"
"5 Industry Portfolios [Daily]"
"10 Industry Portfolios"
"10 Industry Portfolios [ex. Dividends]"
"10 Industry Portfolios [Daily]"
"12 Industry Portfolios"
"12 Industry Portfolios [ex. Dividends]"
"12 Industry Portfolios [Daily]"
"17 Industry Portfolios"
"17 Industry Portfolios [ex. Dividends]"
"17 Industry Portfolios [Daily]"
"30 Industry Portfolios"
"30 Industry Portfolios [ex. Dividends]"
"30 Industry Portfolios [Daily]"
"38 Industry Portfolios"
"38 Industry Portfolios [ex. Dividends]"
"38 Industry Portfolios [Daily]"
"48 Industry Portfolios"
"48 Industry Portfolios [ex. Dividends]"
"48 Industry Portfolios [Daily]"
"49 Industry Portfolios"
"49 Industry Portfolios [ex. Dividends]"
"49 Industry Portfolios [Daily]"
"ME Breakpoints"
"BE/ME Breakpoints"
"Operating Profitability Breakpoints"
"Investment Breakpoints"
"E/P Breakpoints"
"CF/P Breakpoints"
"D/P Breakpoints"
"Prior (2-12) Return Breakpoints"
"Fama/French Developed 3 Factors"
"Fama/French Developed 3 Factors [Daily]"
"Fama/French Developed ex US 3 Factors"
"Fama/French Developed ex US 3 Factors [Daily]"
"Fama/French European 3 Factors"
"Fama/French European 3 Factors [Daily]"
"Fama/French Japanese 3 Factors"
"Fama/French Japanese 3 Factors [Daily]"
"Fama/French Asia Pacific ex Japan 3 Factors"
"Fama/French Asia Pacific ex Japan 3 Factors [Daily]"
"Fama/French North American 3 Factors"
"Fama/French North American 3 Factors [Daily]"
"Fama/French Developed 5 Factors"
"Fama/French Developed 5 Factors [Daily]"
"Fama/French Developed ex US 5 Factors"
"Fama/French Developed ex US 5 Factors [Daily]"
"Fama/French European 5 Factors"
"Fama/French European 5 Factors [Daily]"
"Fama/French Japanese 5 Factors"
"Fama/French Japanese 5 Factors [Daily]"
"Fama/French Asia Pacific ex Japan 5 Factors"
"Fama/French Asia Pacific ex Japan 5 Factors [Daily]"
"Fama/French North American 5 Factors"
"Fama/French North American 5 Factors [Daily]"
"Developed Momentum Factor (Mom)"
"Developed Momentum Factor (Mom) [Daily]"
"Developed ex US Momentum Factor (Mom)"
"Developed ex US Momentum Factor (Mom) [Daily]"
"European Momentum Factor (Mom)"
"European Momentum Factor (Mom) [Daily]"
"Japanese Momentum Factor (Mom)"
"Japanese Momentum Factor (Mom) [Daily]"
"Asia Pacific ex Japan Momentum Factor (Mom)"
"Asia Pacific ex Japan Momentum Factor (Mom) [Daily]"
"North American Momentum Factor (Mom)"
"North American Momentum Factor (Mom) [Daily]"
"6 Developed Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 Developed Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"6 Developed ex US Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 Developed ex US Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"6 European Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 European Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"6 Japanese Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 Japanese Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"6 North American Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 North American Portfolios Formed on Size and Book-to-Market (2 x 3) [Daily]"
"25 Developed Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 Developed Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"25 Developed ex US Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 Developed ex US Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"25 European Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 European Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"25 Japanese Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 Japanese Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"25 North American Portfolios Formed on Size and Book-to-Market (5 x 5)"
"25 North American Portfolios Formed on Size and Book-to-Market (5 x 5) [Daily]"
"6 Developed Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 Developed Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"6 Developed ex US Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 Developed ex US Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"6 European Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 European Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"6 Japanese Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 Japanese Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"6 North American Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 North American Portfolios Formed on Size and Operating Profitability (2 x 3) [Daily]"
"25 Developed Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 Developed Portfolios Formed on Size and Operating Profitability (5 x 5) [Daily]"
"25 Developed ex US Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 Developed ex US Portfolios Formed on Size and Operating Profitability (5 x 5) [Daily]"
"25 European Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 European Portfolios Formed on Size and Operating Profitability (5 x 5) [Daily]"
"25 Japanese Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 Japanese Portfolios Formed on Size and Operating Profitability (5 x 5) [Daily]"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Operating Profitability (5 x 5) [Daily]"
"25 North American Portfolios Formed on Size and Operating Profitability (5 x 5)"
"25 North American Portfolios Formed on Size and Operating Profitability (5 x 5) [Daily]"
"6 Developed Portfolios Formed on Size and Investment (2 x 3)"
"6 Developed Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"6 Developed ex US Portfolios Formed on Size and Investment (2 x 3)"
"6 Developed ex US Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"6 European Portfolios Formed on Size and Investment (2 x 3)"
"6 European Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"6 Japanese Portfolios Formed on Size and Investment (2 x 3)"
"6 Japanese Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Investment (2 x 3)"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"6 North American Portfolios Formed on Size and Investment (2 x 3)"
"6 North American Portfolios Formed on Size and Investment (2 x 3) [Daily]"
"25 Developed Portfolios Formed on Size and Investment (5 x 5)"
"25 Developed Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"25 Developed ex US Portfolios Formed on Size and Investment (5 x 5)"
"25 Developed ex US Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"25 European Portfolios Formed on Size and Investment (5 x 5)"
"25 European Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"25 Japanese Portfolios Formed on Size and Investment (5 x 5)"
"25 Japanese Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Investment (5 x 5)"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"25 North American Portfolios Formed on Size and Investment (5 x 5)"
"25 North American Portfolios Formed on Size and Investment (5 x 5) [Daily]"
"6 Developed Portfolios Formed on Size and Momentum (2 x 3)"
"6 Developed Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"6 Developed ex US Portfolios Formed on Size and Momentum (2 x 3)"
"6 Developed ex US Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"6 European Portfolios Formed on Size and Momentum (2 x 3)"
"6 European Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"6 Japanese Portfolios Formed on Size and Momentum (2 x 3)"
"6 Japanese Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Momentum (2 x 3)"
"6 Asia Pacific ex Japan Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"6 North American Portfolios Formed on Size and Momentum (2 x 3)"
"6 North American Portfolios Formed on Size and Momentum (2 x 3) [Daily]"
"25 Developed Portfolios Formed on Size and Momentum (5 x 5)"
"25 Developed Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"25 Developed ex US Portfolios Formed on Size and Momentum (5 x 5)"
"25 Developed ex US Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"25 European Portfolios Formed on Size and Momentum (5 x 5)"
"25 European Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"25 Japanese Portfolios Formed on Size and Momentum (5 x 5)"
"25 Japanese Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Momentum (5 x 5)"
"25 Asia Pacific ex Japan Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"25 North American Portfolios Formed on Size and Momentum (5 x 5)"
"25 North American Portfolios Formed on Size and Momentum (5 x 5) [Daily]"
"32 Developed Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 Developed ex US Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 European Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 Japanese Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 Asia Pacific ex Japan Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 North American Portfolios Formed on Size, Book-to-Market, and Operating Profitability (2 x 4 x 4)"
"32 Developed Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 Developed ex US Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 European Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 Japanese Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 Asia Pacific ex Japan Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 North American Portfolios Formed on Size, Book-to-Market, and Investment (2 x 4 x 4)"
"32 Developed Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"32 Developed ex US Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"32 European Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"32 Japanese Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"32 Asia Pacific ex Japan Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"32 North American Portfolios Formed on Size, Operating Profitability, and Investment (2 x 4 x 4)"
"Fama/French Emerging 5 Factors"
"Emerging Momentum Factor (Mom)"
"6 Emerging Market Portfolios Formed on Size and Book-to-Market (2 x 3)"
"6 Emerging Market Portfolios Formed on Size and Operating Profitability (2 x 3)"
"6 Emerging Market Portfolios Formed on Size and Investment (2 x 3)"
"6 Emerging Market Portfolios Formed on Size and Momentum (2 x 3)"
"4 Emerging Market Portfolios Formed on Book-to-Market  and Operating Profitability (2 x 2)  "
"4 Emerging Market Portfolios Formed on Operating Profitability and Investment (2 x 2)"
"4 Emerging Market Portfolios Formed on Book-to-Market and Investment (2 x 2)"
}
