% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_fmp_fin}
\alias{get_fmp_fin}
\title{Get Financial Statements from Financial Modeling Prep}
\usage{
get_fmp_fin(ticker, api_key, report = "is", period = "annual")
}
\arguments{
\item{period}{}
}
\description{
Get Financial Statements from Financial Modeling Prep
}
