% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_profiles}
\alias{get_profiles}
\title{Get Company Profile from Financial Modeling Prep}
\usage{
get_profiles(api_key, temp_dir = tempdir(), time_out = 300)
}
\arguments{
\item{time_out}{}
}
\description{
Get Company Profile from Financial Modeling Prep
}
