% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{chart_returnscatter}
\alias{chart_returnscatter}
\title{Return Scatter Plot}
\usage{
chart_returnscatter(
  series1,
  series2,
  data = "prices",
  field = "adjusted",
  name1 = NULL,
  name2 = NULL,
  title = "Return Scatter Plot",
  subtitle = "dates",
  trend = T,
  timescale = "alpha",
  labels = "dates",
  pts.size = 0.1,
  pts.color = "black",
  lm.size = 0.1,
  lm.color = "red"
)
}
\arguments{
\item{timescale}{one of "alpha", "color", or "none"}

\item{labels}{one of "dates" or "none"}

\item{lm.color}{}
}
\description{
Return Scatter Plot
}
